move theme to presentation/common
