+++ lib/presentation/app/my_app.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:j_calc/presentation/app/theme.dart';
 import 'package:j_calc/presentation/common/l10n/app_localizations.dart';
 import 'package:j_calc/presentation/common/router/router.dart';
+import 'package:j_calc/presentation/common/theme/theme.dart';
 import 'package:j_calc/presentation/features/calculator/cubit/calculator_cubit.dart';
 import 'package:j_calc/presentation/features/history/cubit/history_cubit.dart';
